server/pyproject.toml
[tool.poetry]
name = "PersonTracker"
version = "0.1.0"
description = ""
authors = ["Henrique Beckmann <henriquebeckmann2@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
cmake = "^3.26.3"
dlib = "^19.24.1"
numpy = "^1.24.3"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
python-multipart = "^0.0.6"
opencv-python = "^4.7.0.72"
face-recognition = "^1.3.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
